mojupdate

+++ players/python/model.py
@@ -123,8 +123,8 @@ class PPO:
     def update(self, memory: PPOMemory, ghost: ShadeID) -> None:
         boards: Tensor = torch.stack(memory['board'][ghost])
         extras: Tensor = torch.stack(memory['extra'][ghost])
-        actions: Tensor = torch.tensor(memory['actions'][ghost])
-        old_log_probs: Tensor = torch.tensor(memory['log_probs'][ghost])
+        actions: Tensor = torch.tensor(memory['actions'][ghost], dtype=torch.long)
+        old_log_probs: Tensor = torch.tensor(memory['log_probs'][ghost], dtype=torch.float32)
         rewards: List[float] = memory['rewards'][ghost]
         dones: List[bool] = memory['dones'][ghost]
